dyna q fully debugged

+++ src/agents/q_learning.py
@@ -3,7 +3,7 @@ import numpy as np
 
 class QLearningAgent(Agent):
 
-    def step(self, reward:int , new_position:np.ndarray , current_timestep:int) -> None:
+    def step(self, reward:float , new_position:np.ndarray , current_timestep:int) -> None:
         x_0 , x_1 , y_0 , y_1 , _ = self._act(new_position , current_timestep)
         self.q[ y_0 , x_0 , self.action.value] += self.alpha * ( # q(s , a) = q(s , a) + alpha * ( reward + gamma * (q(s' , a') - q(s , a))
             reward + self.discount_rate * (self.q[y_1 , x_1].max()) - self.q[y_0 , x_0 , self.action.value])
